third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-outline-shorthand-expected.txt

Tests that the outline shorthand is computed properly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS computedStyle.getPropertyValue('outline') is 'rgb(255, 0, 0) solid 5px'
PASS computedStyle.getPropertyValue('outline') is 'rgb(0, 0, 0) solid 32px'
PASS computedStyle.getPropertyValue('outline') is 'rgb(255, 0, 0) none 0px'
PASS successfullyParsed is true

TEST COMPLETE
